chore(py/aioia): clean up aioia package
CHANGELOG:
- [ ] Remove the aioia package.
