--- py/packages/aioia/README.md
@@ -1,3 +1,0 @@
-# aioia
-
-Asyncio server management functionality for Genkit.
