--- py/packages/aioia/pyproject.toml
@@ -1,23 +1,0 @@
-[project]
-dependencies = [
-  "asgiref>=3.8.1",
-  "granian>=2.0.1",
-  "psutil>=7.0.0",
-  "structlog>=25.1.0",
-  "types-psutil>=7.0.0.20250218",
-  "typing-extensions>=4.12.2",
-  "uvicorn>=0.34.0",
-  "uvloop>=0.21.0",
-]
-description = "Useful tools for working with asyncio."
-name = "aioia"
-readme = "README.md"
-requires-python = ">=3.11"
-version = "0.1.0"
-
-[build-system]
-build-backend = "hatchling.build"
-requires      = ["hatchling"]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/aioia"]
