+++ py/samples/multi_server/pyproject.toml
@@ -1,12 +1,12 @@
 [project]
 dependencies = [
-  "aioia",
   "genkit-ai",
   "asgiref>=3.8.1",
   "litestar>=2.15.1",
   "starlette>=0.46.1",
   "structlog>=25.2.0",
   "uvloop>=0.21.0",
+  "uvicorn>=0.34.0",
 ]
 description = "Sample implementation to exercise the Genkit multi server manager."
 name = "multi-server"
